+++ contracts/secret-order-book/src/msg.rs
@@ -123,7 +123,7 @@ pub struct LimitOrderState {
     pub is_bid: bool,
     pub status: String, //Active, PartiallyFilled, Filled
     pub price: Uint128,
-    pub order_token_index: usize,
+    pub order_token_index: i8,
     pub order_token_init_quant: Uint128,
     pub balances: Vec<Uint128>,
     pub timestamp: u64
